move_ground CA: avoid corner hexes for workers also
When a planet is threatened, the CA moves units away from corner hexes at the end of the evaluation, with lower priority than the other moves. This could previously lead to units being shuffled back and forth due to this action competing with moving workers to productive terrain. Now corner hexes are also avoided when finding good terrain for workers, if the planet is threatened by enemy transports with passengers.