transport_troops CA: decrease distance for homeworld threats
On the smaller maps, enemy transports are almost always within 2 moves of the AI homeworld, meaning that previously units on the homeworld were almost never available on those maps.  With the recent addition of the defend_homeworld purpose for moving units back to the homeworld when threatened, it should be okay to decrease this distance even for the larger maps.